dsaaa: ch01: Provide all answers for intro exercises
The chapter 1 introduces philosophy of data structures, e.g. the need,
cost and benefits. It also define the abstract data types (ADT), what
is problems, algorithms and programs. Some design patterns are briefly
mentioned such as flyweight, visitor, composite, and strategy. This
commit provides the answers for exercises from 1.1 to 1.20.

Signed-off-by: Hieu Vu <hieutrgvu@gmail.com>