c++htp: ch05: Provide answers for control statement part 2
This commit provides all answers for the exercises from 5.1 to 5.32.
Chapter 5 demonstrates the for, do...while and switch statements. Along
with if, if...else, while, these control statements support both
structured and object-oriented programming. It also introduces the break
and continue statements, the logical operators (&&, ||, !), and the
confusing between == and = operators.

Signed-off-by: Hieu Vu <hieutrgvu@gmail.com>